Fix downloading of books whose title has trailing whitespace
This fixes the "Unchecked lastError value: Error: filename must not
contain illegal characters" error while trying to download an audiobook
whose title has a space at the end
(https://akniga.org/valeriya-volchya-semya-bruts). The generated
filenames had the following pattern:

```
Волчья Валерия - Семья Брутс /cover.jpg
Волчья Валерия - Семья Брутс /desc.txt
```

which are fine (verified in OS X 10.14.6), but Firefox 84.0.2 complained
about them.

The fix is to trim `path` after it's constructed from the `author` and
`title` strings so that it doesn't have leading or trailing whitespace.